requirements.txt
telethon==1.36.0
google-generativeai==0.8.3
python-dotenv==1.0.1
aiohttp==3.10.5
asyncio==3.4.3
APScheduler==3.10.4
Flask==3.0.3
gunicorn==22.0.0
cryptg==0.4.0
Pillow>=10.2.0
